% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{det_info_logistic}
\alias{det_info_logistic}
\title{Determinant of the Fisher information matrix of the
two-parameter logistic regression.}
\usage{
det_info_logistic(x, N_n, alpha, beta)
}
\arguments{
\item{x}{Vector of values (or transformed values; e.g. log) of doses.}

\item{N_n}{Total sample size at both the escalation and the expansion phases.}

\item{alpha}{Intercept of the logistic regression}

\item{beta}{Slope of the logistic regression}
}
\value{
Sufficient range of response rates.
}
\description{
Calculate the determinant of the Fisher information matrix of
the two-parameter logistic regression
}
\examples{
x     = (1:4)*0.1
N_n   = c(6, 3, 3, 6)
alpha = -1.4
beta  = 5.6

det_info_logistic(x, N_n, alpha, beta)

}
