% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_bayes_logistic_model}
\alias{sl_bayes_logistic_model}
\title{Bayesian two-parameter logistic regression model with
normal prior distribution.}
\usage{
sl_bayes_logistic_model(
  stan_data,
  mu_prior = c(0, 0),
  sigma_prior = c(100, 100)
)
}
\arguments{
\item{stan_data}{Data frame including x: vector of values (or
transformed values; e.g. log) of doses;
y: vector of number of responses;
N_n: vector of total number of subjects.}

\item{mu_prior}{Vector of the normal mean for two-parameter
logistic regression.}

\item{sigma_prior}{Vector of the normal standard error for two-parameter
logistic regression.}
}
\value{
post_alpha: posterior sample of intercept; post_beta:
posterior sample of slope; post_q_all: posterior sample of
response rates for all doses.
}
\description{
Fit the Bayesian two-parameter logistic regression
model with normal prior distribution.
}
\examples{
stan_data = data.frame(x   = (1:4)*0.1,
                       y   = c(1, 3, 4, 5),
                       N_n = c(16, 10, 10, 16))
mu_prior    = c(0, 0)
sigma_prior = c(100, 100)

sl_bayes_logistic_model(stan_data, mu_prior, sigma_prior)

}
