% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_data_dede}
\alias{sl_data_dede}
\title{Combined data of the dose-escalation phase data and
the dose-expansion phase data using the skip-levels (SL) design or
the uniformly-assign (UA) design.}
\usage{
sl_data_dede(data_escal, data_expan, dose, x, p, q, gamma)
}
\arguments{
\item{data_escal}{Dose-escalation phase data. Simulated from function
sl_data_escal_sim(). Users can use other
dose-escalation design besides BOIN with the same
structure as the output of function sl_data_escal_sim().}

\item{data_expan}{Dose-expansion phase data. Simulated from function
sl_data_expan_sim().}
}
\value{
data_dede: dose-escalation and dose-expansion phases combined data;
data_dede_DL: the summary of the combined data per DL.
}
\description{
Combine the dose-escalation phase data and
the dose-expansion phase data using the skip-levels (SL) design or
the uniformly-assign (UA) design.
}
\examples{
set.seed(1000)

gamma  = 0.5
p      = c(0.03, 0.07, 0.10, 0.13, 0.25, 0.55)
q      = c(0.20, 0.30, 0.54, 0.57, 0.60, 0.61)
dose   = (1:6)*0.1
x      = dose
p_DLT  = 0.3
n_size = 3
n_max  = 15
n      = 36
skip   = 2
method = "SL"

# escalation data
data_escal = sl_data_escal_sim(p, q, gamma, dose, x, p_DLT, n_size, n_max)
MTD        = sl_select_MTD(data_escal, p_DLT)

# expansion data
rst_expan  = sl_data_expan_sim(p, q, gamma, dose, x, MTD, n, skip, method)
data_expan = rst_expan$data_expan

# combined data
sl_data_dede(data_escal, data_expan, dose, x, p, q, gamma)

}
