% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_data_escal_sim}
\alias{sl_data_escal_sim}
\title{Simulate dose-escalation phase data using BOIN design.}
\usage{
sl_data_escal_sim(
  p,
  q,
  gamma,
  dose,
  x,
  p_DLT = 0.3,
  n_size = 3,
  n_max = 15,
  ...
)
}
\arguments{
\item{p}{Vector of the toxicity rates.}

\item{q}{Vector of the response rates.}

\item{gamma}{Odds ratio of the response rates in toxic group versus that
in non-toxic group.}

\item{dose}{Vector of values of doses.}

\item{x}{Vector of transformed values of doses (e.g. log).}

\item{p_DLT}{Target toxicity rate.}

\item{n_size}{Cohort size.}

\item{n_max}{Maximum sample size per dose.}

\item{...}{Other inputs for BOIN design. See the inputs
from get.boundary().}
}
\value{
Dose-escalation phase data.
}
\description{
Simulate dose-escalation phase data using BOIN design under
multinomial distribution assumption for response and toxicity.
}
\examples{
set.seed(1000)

gamma  = 0.5
p      = c(0.03, 0.07, 0.10, 0.13, 0.25, 0.55)
q      = c(0.20, 0.30, 0.54, 0.57, 0.60, 0.61)
dose   = (1:6)*0.1
x      = dose
p_DLT  = 0.3
n_size = 3
n_max  = 15

sl_data_escal_sim(p, q, gamma, dose, x, p_DLT, n_size, n_max)

}
