% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_data_expan_sim}
\alias{sl_data_expan_sim}
\title{Simulate dose-expansion phase data using skip-levels (SL) design
or uniformly-assign (UA) design.}
\usage{
sl_data_expan_sim(p, q, gamma, dose, x, n, MTD, skip, method = c("SL", "UA"))
}
\arguments{
\item{p}{Vector of the toxicity rates.}

\item{q}{Vector of the response rates.}

\item{gamma}{Odds ratio of the response rates in toxic group versus that
in non-toxic group.}

\item{dose}{Vector of values of doses.}

\item{x}{Vector of transformed values of doses (e.g. log).}

\item{n}{Total number of subjects assigned to the expansion phase.}

\item{MTD}{The DL of the MTD. Get it from function sl_select_MTD().}

\item{skip}{The number of DLs to be skipped.}

\item{method}{Two options: "SL" (the SL design) or "UA" (the UA design).}
}
\value{
data_expan: Dose-expansion phase data; CDR: RP2D candidates region.
}
\description{
Simulate dose-expansion phase data using skip-levels (SL)
design or uniformly-assign (UA) design given the dose level (DL) of the
maximum tolerated dose (MTD) and the number of skip.
}
\examples{
set.seed(1000)

gamma  = 0.5
p      = c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3)
q      = c(0.05, 0.2, 0.35, 0.5, 0.45, 0.4)
dose   = (1:6)*0.1
x      = dose
n      = 30
MTD    = 6
skip   = 1
method = "SL"

sl_data_expan_sim(p, q, gamma, dose, x, n, MTD, skip, method)

}
