% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{sl_data_sim}
\alias{sl_data_sim}
\title{Generate dose-response data.}
\usage{
sl_data_sim(x, q, N_n)
}
\arguments{
\item{x}{Vector of values (or transformed values; e.g. log) of doses.}

\item{q}{Vector of response rates.}

\item{N_n}{Vector of cumulative sample sizes at both escalation and
expansion phases.}
}
\value{
A data frame with three columns: x (values or transformed
values of doses),
n (sample sizes), y (number of responders).
}
\description{
Generate dose-response data given sample size and
response rate per dose.
}
\examples{
set.seed(1000)

x   = (1:4)*0.1
q   = c(0.20, 0.31 ,0.45, 0.60)
N_n = c(16, 10, 10, 16)

sl_data_sim(x, q, N_n)

}
