% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_multinom_par}
\alias{sl_multinom_par}
\title{Parameter values of response-toxicity multinomial distribution.}
\usage{
sl_multinom_par(gamma, p, q)
}
\arguments{
\item{gamma}{Odds ratio of the response rates in toxic group versus
that in non-toxic group.}

\item{p}{Vector of the toxicity rates.}

\item{q}{Vector of the response rates.}
}
\value{
P: the parameters p_11, p_10, p_01, p_00 (11: a toxicity
and a response; 10: a toxicity and no response; 01: no toxicity
and a response; 00: no toxicity and no response); rho: the correlation
of the binary toxicity indicator and the binary response indicator.
}
\description{
Get the parameter values of response-toxicity
multinomial distribution given the toxicity rate, the response rate and
the odds ratio.
}
\examples{
gamma = 0.5
p     = c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3)
q     = c(0.05, 0.2, 0.35, 0.5, 0.45, 0.4)

sl_multinom_par(gamma, p, q)

}
