% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{sl_post_mode_logistic}
\alias{sl_post_mode_logistic}
\title{Posterior mode of parameters of Bayesian two-parameter
logistic regression.}
\usage{
sl_post_mode_logistic(data, mu_prior = c(0, 0), sigma_prior = c(100, 100))
}
\arguments{
\item{data}{Data from function sl_data_sim().}

\item{mu_prior}{Mean of the prior normal distribution.}

\item{sigma_prior}{Standard error of the prior normal distribution.}
}
\value{
Posterior mode of parameters.
}
\description{
Calculate posterior mode of parameters of Bayesian
two-parameter logistic regression.
}
\examples{
x    = (1:4)*0.1
q    = c(0.20, 0.31 ,0.45, 0.60)
N_n  = c(16, 10, 10, 16)
data = sl_data_sim(x, q, N_n)

mu_prior    = c(0, 0)
sigma_prior = c(10, 10)
sl_post_mode_logistic(data, mu_prior, sigma_prior)

}
