% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{sl_q_range_emp_ind}
\alias{sl_q_range_emp_ind}
\title{Independent empirical range of response rates that the skip-levels
(SL) design is more efficient than the uniformly-assign (UA) design
under two-parameter logistic regression.}
\usage{
sl_q_range_emp_ind(x, n, N, by = 0.01)
}
\arguments{
\item{x}{Vector of values (or transformed values; e.g. log) of doses.}

\item{n}{Total sample size at the expansion phase.}

\item{N}{Vector of sample size at each dose level.}

\item{by}{Space between the grid search points.}
}
\value{
Sufficient range of response rates.
}
\description{
Get the grid searched independent empirical range of
response rates of all doses (symmetric at q = 0.5) that
the SL design is more efficient than
the UA design under two-parameter logistic regression in terms of
the determinant of the Fisher information.
}
\examples{
x  = (1:4)*0.1
n  = 28
N  = c(3, 3, 3, 6)
by = 0.01

sl_q_range_emp_ind(x, n, N, by)

}
