% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{sl_q_range_es}
\alias{sl_q_range_es}
\title{Sufficient range of response rates that the skip-levels (SL) design is
more efficient than the uniformly-assign (UA) design under two-parameter
logistic regression when the values of doses (or transformed doses) are
equally spaced.}
\usage{
sl_q_range_es(n = 28, N = rep(0, 4))
}
\arguments{
\item{n}{Total sample size at the expansion phase.}

\item{N}{Vector of sample size at each dose level.}
}
\value{
Sufficient range of response rates.
}
\description{
Get the sufficient range of response rates of all doses that
the SL design is more efficient than the UA design under two-parameter
logistic regression in terms of the determinant of the Fisher information
when the values of doses (or transformed doses) are equally spaced.
}
\examples{
n    = 28
N    = c(3, 3, 3, 6)
sl_q_range_es(n, N)

}
