% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{sl_q_range_plot}
\alias{sl_q_range_plot}
\title{Region plot for sufficient region of the theorem, empirical region
and independent empirical region that the SL design is more efficient than
the UA design under two-parameter logistic regression in terms of
the determinant of the Fisher information.}
\usage{
sl_q_range_plot(x, n, N, by = 0.01)
}
\arguments{
\item{x}{Vector of values (or transformed values; e.g. log) of doses.}

\item{n}{Total sample size at the expansion phase.}

\item{N}{Vector of sample size at each dose level.}

\item{by}{Space between the grid search points.}
}
\value{
Sufficient range of response rates.
}
\description{
Compare the regions of the theorem, empirical grid search and
the empirical grid search with similar structure as that of the theorem that
the SL design is more efficient than
the UA design under two-parameter logistic regression in terms of
the determinant of the Fisher information.
}
\examples{
x  = (1:4)*0.1
n  = 28
N  = c(3, 3, 3, 6)
by = 0.01

sl_q_range_plot(x, n, N, by)

}
