% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_scenario_dede}
\alias{sl_scenario_dede}
\title{Scenarios of dose-response curves under logistic regression.}
\usage{
sl_scenario_dede(max_skip = 2, dose = 1:6, q_MTD = 0.6, MTD = 5, delta = 0.9)
}
\arguments{
\item{max_skip}{Maximum DLs to be skipped. The results will show the case
when skipping 0 to max_skip DLs.}

\item{dose}{Vector of values of doses.}

\item{q_MTD}{Response rate at the MTD.}

\item{MTD}{The DL of the MTD.}

\item{delta}{Relative efficiency of the response rates of the
non-inferiority dose and the MTD. It is a
constant in (0, 1).}
}
\value{
q_all: response rates matrix, each row represents the response
rates scenario for each skip value; par_all: parameters matrix, each row
represents the values of parameters per each skip value.
}
\description{
Get the dose-response curves under logistic regression, where
the curve is determined by the dose-response data point of the maximum
tolerated dose (MTD) and the MTD - (k+1) with response rate equals
delta*q_MTD (k: number of dose levels (DLs) to be skipped; q_MTD:
respnose rate at MTD; delta: a large constant in (0, 1)).
}
\examples{
sl_scenario_dede(max_skip = 2, 
                 dose     = (1:6)*0.1, 
                 q_MTD    = 0.6, 
                 MTD      = 5, 
                 delta    = 0.9)

}
