% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_fun.R
\name{sl_scenario_range}
\alias{sl_scenario_range}
\title{Get the two-parameter logistic regression given the ranges
of the response rates.}
\usage{
sl_scenario_range(
  dose = (1:4) * 0.1,
  q_range = rbind(c(0.2, 0.6), c(0.3, 0.7))
)
}
\arguments{
\item{dose}{Vector of values of doses.}

\item{q_range}{Matrix of ranges of the response rates q (a matrix with
2 columns, each row represents the range of q).}
}
\value{
The list of the response rates and the values of the parameters.
}
\description{
Get the response rates and the values of the parameters
under logistic regression given the ranges of the response rates.
}
\examples{
dose    = (1:4)*0.1
q_range = rbind(c(0.2, 0.6), c(0.3, 0.7), c(0.4, 0.8))
sl_scenario_range(dose, q_range)

}
