% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_select_MTD}
\alias{sl_select_MTD}
\title{Select maximum tolerated dose (MTD).}
\usage{
sl_select_MTD(data_escal, p_DLT = 0.3)
}
\arguments{
\item{data_escal}{Dose-escalation data from function sl_data_escal_sim().}

\item{p_DLT}{Target toxicity rate, which should be the same as the
input in function sl_data_escal_sim().}
}
\value{
Selected dose level (DL) of the MTD.
}
\description{
Select MTD give the dose-escalation data.
}
\examples{
set.seed(1000)

gamma  = 0.5
p      = c(0.03, 0.07, 0.10, 0.13, 0.25, 0.55)
q      = c(0.20, 0.30, 0.54, 0.57, 0.60, 0.61)
dose   = (1:6)*0.1
x      = dose
p_DLT  = 0.3
n_size = 3
n_max  = 15

data_escal = sl_data_escal_sim(p, q, gamma, dose, x, p_DLT, n_size, n_max)

sl_select_MTD(data_escal, p_DLT)

}
