% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_ss_assign_fun}
\alias{sl_ss_assign_fun}
\title{Uniformly assign subjects to each dose level (DL).}
\usage{
sl_ss_assign_fun(n, m)
}
\arguments{
\item{n}{Total number of subjects.}

\item{m}{Number of DLs.}
}
\value{
Number of subjects assigned to each DL.
}
\description{
Get the number of subjects assigned to each DL uniformly.
}
\examples{
sl_ss_assign_fun(30, 4)

}
