% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_bayes_logistic_model_dede}
\alias{sl_bayes_logistic_model_dede}
\title{Bayesian two-parameter logistic regression model with
normal prior distribution given the combined dose-escalation
and dose-expansion phases data and the RP2D candidates region (CDR).}
\usage{
sl_bayes_logistic_model_dede(
  data_dede_DL,
  CDR,
  mu_prior = c(0, 0),
  sigma_prior = c(100, 100)
)
}
\arguments{
\item{data_dede_DL}{Data from function sl_data_dede().}

\item{CDR}{RP2D candidates region; i.e.
DLs within \link{MTD-(skip+1), MTD}.}

\item{mu_prior}{Vector of the normal mean for two-parameter
logistic regression.}

\item{sigma_prior}{Vector of the normal standard error for two-parameter
logistic regression.}
}
\value{
post_alpha: posterior sample of intercept; post_beta:
posterior sample of slope; post_q_all: posterior sample of
response rates for all doses.
}
\description{
Fit the Bayesian two-parameter logistic regression model with
normal prior distribution given the combined dose-escalation
and dose-expansion phases data and the CDR. The data with DLs within CDR
will be used to build the model.
}
\examples{
set.seed(1000)

gamma       = 0.5
p           = c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3)
q           = c(0.05, 0.2, 0.35, 0.5, 0.45, 0.4)
dose        = (1:6)*0.1
x           = dose
p_DLT       = 0.3
n_size      = 3
n_max       = 15
n           = 36
skip        = 2
method      = "SL"
mu_prior    = c(0, 0)
sigma_prior = c(100, 100)

# escalation data
data_escal = sl_data_escal_sim(p, q, gamma, dose, x, p_DLT, n_size, n_max)
MTD        = sl_select_MTD(data_escal, p_DLT)

# expansion data
rst_expan  = sl_data_expan_sim(p, q, gamma, dose, x, MTD, n, skip, method)
data_expan = rst_expan$data_expan
CDR        = rst_expan$CDR

# combined data
rst_data_dede = sl_data_dede(data_escal, data_expan, dose, x, p, q, gamma)
data_dede_DL  = rst_data_dede$data_dede_DL

# model
sl_bayes_logistic_model_dede(data_dede_DL, CDR, mu_prior, sigma_prior)

}
