% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_dede_full_sim}
\alias{sl_dede_full_sim}
\title{Simulation for dose-escalation and dose-expansion phases.}
\usage{
sl_dede_full_sim(
  p,
  q,
  gamma,
  dose,
  x,
  p_DLT = 0.3,
  n_size = 3,
  n_max = 15,
  n = 36,
  skip = 2,
  method = c("SL", "UA"),
  mu_prior = c(0, 0),
  sigma_prior = c(100, 100),
  delta = 0.9,
  cc = c(0.7, 0.8),
  ...
)
}
\arguments{
\item{p}{Vector of the toxicity rates.}

\item{q}{Vector of the response rates.}

\item{gamma}{Odds ratio of the response rates in toxic group
versus that in non-toxic group.}

\item{dose}{Vector of values of doses.}

\item{x}{Vector of transformed values of doses (e.g. log).}

\item{p_DLT}{Target toxicity rate.}

\item{n_size}{Cohort size.}

\item{n_max}{Maximum sample size per dose.}

\item{n}{Total number of subjects assigned to the expansion phase.}

\item{skip}{The number of DLs to be skipped.}

\item{method}{Two options: "SL" (the SL design) or "UA"
(the UA design).}

\item{mu_prior}{Vector of the normal mean for two-parameter logistic
regression.}

\item{sigma_prior}{Vector of the normal standard error for two-parameter
logistic regression.}

\item{delta}{Relative efficiency of the response rates of the
non-inferiority dose and the MTD. It is a constant
in (0, 1).}

\item{cc}{Posterior probability cut to decide if the low dose is
non-inferior to the MTD. It is a constant in (0, 1).}

\item{...}{Other inputs for BOIN design. See the inputs
from get.boundary().}
}
\value{
data_dede: combined dose-escalation and dose-expansion data;
data_dede_DL: summary of the combined data; post_par: posterior sample of
intercept, slope and the response rates; rst_AR: the AR for each cc.
}
\description{
Simulate dose-escalation and dose-expansion phases data and
calculate the maximum tolerated dose (MTD) and the admissible region
(AR) for RP2Ds.
}
\examples{
set.seed(1000)

gamma       = 0.5
p           = c(0.03, 0.07, 0.10, 0.13, 0.25, 0.55)
q           = c(0.20, 0.30, 0.54, 0.57, 0.60, 0.61)
dose        = (1:6)*0.1
x           = dose
p_DLT       = 0.3
n_size      = 3
n_max       = 15
n           = 36
skip        = 2
method      = "SL"
mu_prior    = c(0, 0)
sigma_prior = c(100, 100)
delta       = 0.9
cc          = c(0.7, 0.8)

sl_dede_full_sim(p, q, gamma, dose, x, 
                 p_DLT, n_size, n_max, 
                 n, skip, method, 
                 mu_prior, sigma_prior, 
                 delta, cc)

}
