% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_sim.R
\name{sl_get_ARLD}
\alias{sl_get_ARLD}
\title{Get the lowest dose level (DL) of the RP2D admissible region (AR).}
\usage{
sl_get_ARLD(post_q_all, CDR, delta = 0.9, cc = 0.7)
}
\arguments{
\item{post_q_all}{Posterior sample of response rates for all DLs from
function sl_bayes_logistic_model_dede().}

\item{CDR}{RP2D candidates region; i.e.
DLs within \link{MTD-(skip+1), MTD}.}

\item{delta}{Relative efficiency of the response rates of the
non-inferiority dose and the MTD. It is a
constant in (0, 1).}

\item{cc}{Posterior probability cut to decide if the low dose is
non-inferior to the MTD. It is a constant in (0, 1).}
}
\value{
The lowest DL of the AR.
}
\description{
Get the lowest dose level (DL) of the RP2D
admissible region (AR) using the Bayesian two-parameter logistic
regression.
}
\examples{
set.seed(1000)

gamma       = 0.5
p           = c(0.03, 0.07, 0.10, 0.13, 0.25, 0.55)
q           = c(0.20, 0.30, 0.54, 0.57, 0.60, 0.61)
dose        = (1:6)*0.1
x           = dose
p_DLT       = 0.3
n_size      = 3
n_max       = 15
n           = 36
skip        = 2
method      = "SL"
mu_prior    = c(0, 0)
sigma_prior = c(100, 100)
delta       = 0.9
cc          = 0.7

# escalation data
data_escal = sl_data_escal_sim(p, q, gamma, dose, x, p_DLT, n_size, n_max)
MTD        = sl_select_MTD(data_escal, p_DLT)

# expansion data
rst_expan  = sl_data_expan_sim(p, q, gamma, dose, x, MTD, n, skip, method)
data_expan = rst_expan$data_expan
CDR        = rst_expan$CDR

# combined data
rst_data_dede = sl_data_dede(data_escal, data_expan, dose, x, p, q, gamma)
data_dede_DL  = rst_data_dede$data_dede_DL

# model
post_par = sl_bayes_logistic_model_dede(data_dede_DL, CDR, 
                                        mu_prior, sigma_prior)

# The lowest DL of the AR
sl_get_ARLD(post_par$post_q_all, CDR, delta, cc)

}
