% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldesign_packages.R
\docType{package}
\name{sldesign_packages}
\alias{sldesign-package}
\alias{sldesign_packages}
\alias{sldesign}
\title{The 'sldesign' package for skip-Levels design for phase I oncology study}
\description{
A tool to calculate the sufficient range of the response rates
that the skip-levels  (SL) design performs better than the uniformly-assign
(UA) design, as well as conduct simulation studies for the phase 1
dose-escalation and dose-expansion using the SL design or the UA design
and find the recommended phase 2 dose admissible region using Bayesian
2-parameter logistic regression.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan.
R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Weibin Zhong \email{wzhong3@gmu.edu}

Authors:
\itemize{
  \item Weibin Zhong \email{wzhong3@gmu.edu}
  \item Jing-ou Liu
  \item chenguang Wang
}

}
